Renamed Package from General to Generic
