+++ src/main/java/com/beanbeanjuice/main/BeanBot.java
@@ -3,8 +3,8 @@ package com.beanbeanjuice.main;
 import com.beanbeanjuice.command.fun.MemeCommand;
 import com.beanbeanjuice.command.fun.JokeCommand;
 import com.beanbeanjuice.command.fun.SetCountingChannelCommand;
-import com.beanbeanjuice.command.general.HelpCommand;
-import com.beanbeanjuice.command.general.PingCommand;
+import com.beanbeanjuice.command.generic.HelpCommand;
+import com.beanbeanjuice.command.generic.PingCommand;
 import com.beanbeanjuice.command.moderation.*;
 import com.beanbeanjuice.command.moderation.mute.MuteCommand;
 import com.beanbeanjuice.command.moderation.mute.UnMuteCommand;
